refactor(editor): Reduce PR scope by dropping shared balance helper and test file split
