+++ packages/frontend/editor-ui/src/app/composables/useSettingsItems.ts
@@ -8,7 +8,6 @@ import { VIEWS } from '../constants';
 import { useUIStore } from '../stores/ui.store';
 import { useSettingsStore } from '../stores/settings.store';
 import { hasPermission } from '../utils/rbac/permissions';
-import { formatWalletBalance } from '../utils/aiGatewayUtils';
 import { MIGRATION_REPORT_TARGET_VERSION } from '@n8n/api-types';
 import { useEnvFeatureFlag } from '@/features/shared/envFeatureFlag/useEnvFeatureFlag';
 
@@ -64,7 +63,12 @@ export function useSettingsItems() {
 				available:
 					settingsStore.isAiGatewayEnabled && canUserAccessRouteByName(VIEWS.AI_GATEWAY_SETTINGS),
 				route: { to: { name: VIEWS.AI_GATEWAY_SETTINGS } },
-				creditsTag: formatWalletBalance(balance.value),
+				creditsTag:
+					balance.value !== undefined
+						? i18n.baseText('aiGateway.wallet.balanceRemaining', {
+								interpolate: { balance: `$${Number(balance.value).toFixed(2)}` },
+							})
+						: undefined,
 			},
 			{
 				id: 'settings-roles',
