+++ packages/frontend/editor-ui/src/features/shared/nodeCreator/components/SectionHeaderCreditsTag.vue
@@ -1,14 +1,23 @@
 <script setup lang="ts">
 import { computed, onMounted } from 'vue';
 import { N8nTag } from '@n8n/design-system';
+import { useI18n } from '@n8n/i18n';
 import { useAiGatewayStore } from '@/app/stores/aiGateway.store';
 import { useSettingsStore } from '@/app/stores/settings.store';
-import { formatWalletBalance } from '@/app/utils/aiGatewayUtils';
 
 const aiGatewayStore = useAiGatewayStore();
 const settingsStore = useSettingsStore();
+const i18n = useI18n();
 
-const text = computed(() => formatWalletBalance(aiGatewayStore.balance));
+const text = computed(() => {
+	const balance = aiGatewayStore.balance;
+	if (balance === undefined) return undefined;
+	return balance <= 0
+		? i18n.baseText('aiGateway.wallet.noCredits')
+		: i18n.baseText('aiGateway.wallet.balanceRemaining', {
+				interpolate: { balance: `$${balance.toFixed(2)}` },
+			});
+});
 
 onMounted(() => {
 	if (settingsStore.isAiGatewayEnabled) {
